[level 1] Title: 두 정수 사이의 합, Time: 15.00 ms, Memory: 82 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -1,8 +1,11 @@
+import java.util.*;
+
 class Solution {
     public long solution(int a, int b) {
         long answer = 0;
-        for(int i = Math.min(a, b); i <= Math.max(a, b); i++){
-            answer +=i;
+        
+        for(int i=Math.min(a,b); i<=Math.max(a,b); i++){
+            answer += (long) i;
         }
         return answer;
     }
